requirements.txt
asgiref==3.8.1
atomicwrites==1.4.1
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==2.0.12
colorama==0.4.6
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==4.2.18
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
idna==3.10
iniconfig==2.0.0
isort==5.13.2
oauthlib==3.2.2
packaging==24.2
Pillow==9.3.0
pluggy==0.13.1
py==1.11.0
pycparser==2.22
PyJWT==2.10.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2024.2
requests==2.26.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
toml==0.10.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
